requirements.txt
# Core dependencies
streamlit==1.28.0
playwright==1.40.0
selenium==4.15.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
cryptography==41.0.4
python-multipart==0.0.6

# Data processing
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0

# Database and caching
redis==5.0.0
celery==5.3.1
pymongo==4.5.0

# LLM providers
openai==1.3.0
anthropic==0.7.0
transformers==4.34.0
torch==2.1.0
langchain==0.1.0
langchain-community==0.0.10

# RAG and vector databases
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2
tiktoken==0.5.1

# Document processing
pypdf2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3

# Web scraping
requests==2.31.0
aiohttp==3.8.5
selenium-wire==5.1.0
undetected-chromedriver==3.5.3

# API and validation
pydantic==2.4.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0

# Monitoring and utilities
psutil==5.9.5
prometheus-client==0.18.0
python-json-logger==2.0.7
structlog==23.2.0

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78

# Development tools
python-dateutil==2.8.2
pytz==2023.3
